task/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-emit-workspace --no-dev --no-annotate --no-hashes --output-file requirements.txt
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
botocore==1.37.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6 ; sys_platform == 'win32'
debugpy==1.8.14
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
email-validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.7
frozenlist==1.6.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
ijson==3.4.0
inspect-ai==0.3.95
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
markupsafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mmh3==5.1.0
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.82.0
platformdirs==4.3.8
ply==3.11
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.10.2
pydantic-core==2.27.1
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
s3fs==2025.3.0
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.41.3
tenacity==9.1.2
textual==3.2.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.13.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.31.0
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
